start implementing the parser
i will document this before it gets merged. plan is to house all the information in a single file then feed it to the specific languages' build systems \(either manually by creating a temporary file with some prelude definitions for all the info e.g. current LDFLAGS macro in nob\) or giving it via cmd args or something